Move the register allocator specific code from asm_generation/mod.rs into register_allocator.rs.
